Create README.md
Here is a project on Parkinson disease detection. Exploratory data analysis like correlation and plot was used on patient's attribute information to predict the status of each patient. 

With the feature importance derived in this project, doctors, medical practitioners/professionals can know which information to pay more attention to for a more accurate diagnosis.